Removes dependency on \Magento\Store\Model\Store
The Store model requires the session, which require the app state to be defined.

When an error or exception occurs before the app state has been initialised this causes an error that crashes Magento.

The workaround is to use the runMode and scopeCode parameters, which originate from $_SERVER which is where the store code is set.